+++ Oem/ACE/ACE_wrappers/ace/ACE.cpp
@@ -707,7 +707,7 @@ ACE::recvmsg (ACE_HANDLE handle,
         return -1;
       else
         {
-          int bytes_transferred = ACE_OS::recvmsg (handle, msg, flags);
+          ssize_t bytes_transferred = ACE_OS::recvmsg (handle, msg, flags);
           ACE::restore_non_blocking_mode (handle, val);
           return bytes_transferred;
         }
@@ -739,7 +739,8 @@ ACE::recvfrom (ACE_HANDLE handle,
         return -1;
       else
         {
-          int bytes_transferred = ACE_OS::recvfrom (handle, buf, len, flags, addr, addrlen);
+          ssize_t bytes_transferred =
+            ACE_OS::recvfrom (handle, buf, len, flags, addr, addrlen);
           ACE::restore_non_blocking_mode (handle, val);
           return bytes_transferred;
         }
@@ -1518,7 +1519,7 @@ ACE::sendmsg (ACE_HANDLE handle,
         return -1;
       else
         {
-          int bytes_transferred = ACE_OS::sendmsg (handle, msg, flags);
+          ssize_t bytes_transferred = ACE_OS::sendmsg (handle, msg, flags);
           ACE::restore_non_blocking_mode (handle, val);
           return bytes_transferred;
         }
@@ -1550,7 +1551,8 @@ ACE::sendto (ACE_HANDLE handle,
         return -1;
       else
         {
-          int bytes_transferred = ACE_OS::sendto (handle, buf, len, flags, addr, addrlen);
+          ssize_t bytes_transferred =
+            ACE_OS::sendto (handle, buf, len, flags, addr, addrlen);
           ACE::restore_non_blocking_mode (handle, val);
           return bytes_transferred;
         }
