+++ Oem/dbxml-2.3.10/xqilla/src/ast/XQFunction.cpp
@@ -66,9 +66,10 @@ const XMLCh XQFunction::XMLChFunctionURI[] =
     XERCES_CPP_NAMESPACE_QUALIFIER chLatin_n,       XERCES_CPP_NAMESPACE_QUALIFIER chLatin_s,       XERCES_CPP_NAMESPACE_QUALIFIER chNull
 };
 
-const unsigned int XQFunction::UNLIMITED = 10000; // A reasonably large number
+const size_t XQFunction::UNLIMITED = 10000; // A reasonably large number
 
-XQFunction::XQFunction(const XMLCh* name, unsigned int argsFrom, unsigned int argsTo, const char* paramDecl, const VectorOfASTNodes &args, XPath2MemoryManager* memMgr)
+XQFunction::XQFunction(const XMLCh* name, size_t argsFrom, size_t argsTo, const char* paramDecl,
+                       const VectorOfASTNodes &args, XPath2MemoryManager* memMgr)
   : ASTNodeImpl(memMgr),
   _fName(name), _fURI(XMLChFunctionURI), 
   _nArgsFrom(argsFrom), _nArgsTo(argsTo),
@@ -145,7 +146,7 @@ const XMLCh* XQFunction::getFunctionSignature() const
 
 ASTNode *XQFunction::resolveArguments(StaticContext *context, bool checkTimezone, bool numericFunction)
 {
-  unsigned int paramNumber = 0;
+  size_t paramNumber = 0;
   for(VectorOfASTNodes::iterator i = _args.begin(); i != _args.end(); ++i) {
     *i = _paramDecl[paramNumber]->convertFunctionArg(*i, context, numericFunction, this);
     *i = (*i)->staticResolution(context);
@@ -175,14 +176,14 @@ ASTNode *XQFunction::calculateSRCForArguments(StaticContext *context, bool check
   return this;
 }
 
-Result XQFunction::getParamNumber(unsigned int number, DynamicContext* context, int flags) const
+Result XQFunction::getParamNumber(size_t number, DynamicContext* context, int flags) const
 {
 	assert(number > 0);
 	assert(number <= getNumArgs());
 	return _args[number - 1]->collapseTree(context, flags);
 }
 
-unsigned int XQFunction::getNumArgs() const
+size_t XQFunction::getNumArgs() const
 {
   return _args.size();
 }
